Database.Relational.Query.Derives

Description

This module defines typed SQLs derived from type informations.

Query derivation

specifiedKey

Arguments

:: PersistableWidth p

=> Pi a p
Unique key proof object which record type is a and key type is p.

-> Relation () a
Relation to add restriction.

-> Relation p a
Result restricted Relation

Query restricted with specified key.

unique

Arguments

:: PersistableWidth p

=> Key Unique a p
Unique key proof object which record type is a and key type is p.

-> Relation () a
Relation to add restriction.

-> Relation p a
Result restricted Relation

Query restricted with specified unique key.

primary'

Arguments

:: PersistableWidth p

=> Key Primary a p
Primary key proof object which record type is a and key type is p.

-> Relation () a
Relation to add restriction.

-> Relation p a
Result restricted Relation

Query restricted with specified primary key.

primary

Arguments

:: HasConstraintKey Primary a p

=> Relation () a
Relation to add restriction.

-> Relation p a
Result restricted Relation

Query restricted with inferred primary key.

Update derivation

updateByConstraintKey

Arguments

:: Table r
Table to update

-> Key c r p
Key with constraint c, record type r and columns type p

-> KeyUpdate p r
Result typed Update

Typed KeyUpdate using specified constraint key.

primaryUpdate

Arguments

:: HasConstraintKey Primary r p

=> Table r
Table to update

-> KeyUpdate p r
Result typed Update

Typed KeyUpdate using inferred primary key.

updateValuesWithKey :: ToSql q r => Pi r p -> r -> [q]

Convert from Haskell type r into SQL value q list expected by update form like

UPDATE table SET c0 = ?, c1 = ?, ..., cn = ? WHERE key0 = ? AND key1 = ? AND key2 = ? ...

using derived RecordToSql proof object.

Derived objects from table

derivedUniqueRelation

Arguments

:: TableDerivable r

=> Key Unique r k
Unique key proof object which record type is a and key type is p.

-> Projection c k
Unique key value to specify.

-> UniqueRelation () c r
Result restricted Relation

UniqueRelation inferred from table.