Database.Relational.Query.Monad.Assign

Description

This module contains definitions about restrict context with assignment monad type.

Monad to restrict target records with assignment.

type Assign r = Assignings r Restrict

Target update monad type used from update statement and merge statement.

type AssignStatement r a = Projection Flat r -> Assign r a

AssignStatement type synonym. Specifying assignments and restrictions like update statement. Projection record type must be the same as Target type parameter r.

extract :: Assign r a -> Config -> ((a, Table r -> [Assignment]), QueryRestriction Flat)

Run Assign.