Database.Relational.Query.Monad.BaseType

Description

This module defines base monad type to build queries.

Base monad type to build queries

type ConfigureQuery = Qualify (QueryConfig Identity)

Thin monad type for untyped structure.

configureQuery :: ConfigureQuery q -> Config -> q

Run ConfigureQuery monad with initial state to get only result.

qualifyQuery :: a -> ConfigureQuery (Qualified a)

Get qualifyed table form query.

askConfig :: ConfigureQuery Config

Read configuration.

Relation type

data Relation p r

Relation type with place-holder parameter p and query result type r.

unsafeTypeRelation :: ConfigureQuery SubQuery -> Relation p r

Unsafely type qualified subquery into record typed relation type.

untypeRelation :: Relation p r -> ConfigureQuery SubQuery

Sub-query Qualify monad from relation.

relationWidth :: Relation p r -> PersistableRecordWidth r

PersistableRecordWidth of Relation type.

dump :: Relation p r -> String

Dump internal structure tree.

sqlFromRelationWith :: Relation p r -> Config -> StringSQL

Generate SQL string from Relation with configuration.

sqlFromRelation :: Relation p r -> StringSQL

SQL string from Relation.

rightPh :: Relation ((), p) r -> Relation p r

Simplify placeholder type applying left identity element.

leftPh :: Relation (p, ()) r -> Relation p r

Simplify placeholder type applying right identity element.