Database.Relational.Query.Monad.Restrict

Description

This module contains definitions about simple restrict context monad type.

Monad to restrict target records.

type Restrict = Restrictings Flat ConfigureQuery

Restrict only monad type used from update statement and delete statement.

type RestrictedStatement r a = Projection Flat r -> Restrict a

RestrictedStatement type synonym. Projection record type r must be the same as Restrictings type parameter r.

extract :: Restrict a -> Config -> (a, QueryRestriction Flat)

Run Restrict to get QueryRestriction.