Database.Relational.Query.Monad.Trans.Aggregating

Description

This module defines monad transformer which lift from MonadQuery into Aggregated query.

Transformer into aggregated query

data Aggregatings ac at m a

Type to accumulate aggregating context. Type ac is aggregating-context type like aggregating key set building, aggregating key sets set building and partition key set building. Type at is aggregating term type.

aggregatings :: Monad m => m a -> Aggregatings ac at m a

Lift to Aggregatings.

type AggregatingSetT = Aggregatings Set AggregateElem

Context type building one grouping set.

type AggregatingSetListT = Aggregatings SetList AggregateSet

Context type building grouping sets list.

type AggregatingPowerSetT = Aggregatings Power AggregateBitKey

Context type building power group set.

type PartitioningSetT c = Aggregatings c AggregateColumnRef

Context type building partition keys set.

Result

extractAggregateTerms :: (Monad m, Functor m) => Aggregatings ac at m a -> m (a, [at])

Run Aggregatings to get terms list.

Grouping sets support

type AggregatingSet = AggregatingSetT Identity

Context monad type to build single grouping set.

type AggregatingPowerSet = AggregatingPowerSetT Identity

Context monad type to build grouping power set.

type AggregatingSetList = AggregatingSetListT Identity

Context monad type to build grouping set list.

type PartitioningSet c = PartitioningSetT c Identity

Context monad type to build partition keys set.

key :: Projection Flat r -> AggregatingSet (Projection Aggregated (Maybe r))

Specify key of single grouping set from Projection.

key' :: AggregateKey a -> AggregatingSet a

Specify key of single grouping set.

set :: AggregatingSet a -> AggregatingSetList a

Finalize and specify single grouping set.

bkey :: Projection Flat r -> AggregatingPowerSet (Projection Aggregated (Maybe r))

Specify key of rollup and cube power set.

rollup :: AggregatingPowerSet a -> AggregateKey a

Finalize grouping power set as rollup power set.

cube :: AggregatingPowerSet a -> AggregateKey a

Finalize grouping power set as cube power set.

groupingSets :: AggregatingSetList a -> AggregateKey a

Finalize grouping set list.