Database.Relational.Query.Monad.Trans.Restricting

Description

This module defines monad transformer which lift to basic MonadQuery.

Transformer into restricted context

data Restrictings c m a

Type to accumulate query restrictions. Type c is context tag of restriction building like Flat (where) or Aggregated (having).

restrictings :: Monad m => m a -> Restrictings c m a

Lift to Restrictings

Result

extractRestrict :: (Monad m, Functor m) => Restrictings c m a -> m (a, QueryRestriction c)

Run Restrictings to get QueryRestriction