Database.Relational.Query.ProjectableClass

Description

This module provides interfaces to preserve constraints of direct product projections.

Interface to specify record constructors.

class ProductConstructor r where

Specify tuple like record constructors which are allowed to define ProjectableFunctor.

Minimal complete definition

productConstructor

Methods

productConstructor :: r

The constructor which has type r.

ProjectableFunctor and ProjectableApplicative

class ProjectableFunctor p where

Weaken functor on projections.

Minimal complete definition

(|$|)

Methods

(|$|) :: ProductConstructor (a -> b) => (a -> b) -> p a -> p b infixl 4

Method like fmap.

class ProjectableFunctor p => ProjectableApplicative p where

Weaken applicative functor on projections.

Minimal complete definition

(|*|)

Methods

(|*|) :: p (a -> b) -> p a -> p b infixl 4

Method like <*>.

ipfmap :: (ProjectableFunctor p, ProductConstructor (a -> b)) => p a -> p b

Same as |$| other than using inferred record constructor.

Literal SQL terms

class ShowConstantTermsSQL a where

ShowConstantTermsSQL a is implicit rule to derive function to convert from haskell record type a into constant SQL terms.

Generic programming (https://downloads.haskell.org/~ghc/latest/docs/html/users_guide/glasgow_exts.html#generic-programming) with default signature is available for ShowConstantTermsSQL class, so you can make instance like below:

  {-# LANGUAGE DeriveGeneric #-}
  import GHC.Generics (Generic)
  --
  data Foo = Foo { ... } deriving Generic
  instance ShowConstantTermsSQL Foo

Methods

showConstantTermsSQL' :: a -> DList StringSQL

showConstantTermsSQL' :: (Generic a, GShowConstantTermsSQL (Rep a)) => a -> DList StringSQL

showConstantTermsSQL :: ShowConstantTermsSQL a => a -> [StringSQL]

Convert from haskell record to SQL terms list.

type StringSQL = Keyword

String wrap type for SQL strings.