Database.Relational.Query.Pure

Description

This module defines interfaces between haskell pure values and query internal projection values.

Orphan instances

ShowConstantTermsSQL Bool
Constant SQL terms of Bool.

Methods

showConstantTermsSQL' :: Bool -> DList StringSQL

ShowConstantTermsSQL Char
Constant SQL terms of Char.

Methods

showConstantTermsSQL' :: Char -> DList StringSQL

ShowConstantTermsSQL Double
Constant SQL terms of Double. Caution for floating-point error rate.

Methods

showConstantTermsSQL' :: Double -> DList StringSQL

ShowConstantTermsSQL Float
Constant SQL terms of Float. Caution for floating-point error rate.

Methods

showConstantTermsSQL' :: Float -> DList StringSQL

ShowConstantTermsSQL Int
Constant SQL terms of Int. Use this carefully, because this is architecture dependent size of integer type.

Methods

showConstantTermsSQL' :: Int -> DList StringSQL

ShowConstantTermsSQL Int8
Constant SQL terms of Int8.

Methods

showConstantTermsSQL' :: Int8 -> DList StringSQL

ShowConstantTermsSQL Int16
Constant SQL terms of Int16.

Methods

showConstantTermsSQL' :: Int16 -> DList StringSQL

ShowConstantTermsSQL Int32
Constant SQL terms of Int32.

Methods

showConstantTermsSQL' :: Int32 -> DList StringSQL

ShowConstantTermsSQL Int64
Constant SQL terms of Int64.

Methods

showConstantTermsSQL' :: Int64 -> DList StringSQL

ShowConstantTermsSQL ()
Constant SQL terms of '()'.

Methods

showConstantTermsSQL' :: () -> DList StringSQL

ShowConstantTermsSQL String
Constant SQL terms of String.

Methods

showConstantTermsSQL' :: String -> DList StringSQL

ShowConstantTermsSQL ByteString
Constant SQL terms of ByteString.

Methods

showConstantTermsSQL' :: ByteString -> DList StringSQL

ShowConstantTermsSQL ByteString
Constant SQL terms of ByteString.

Methods

showConstantTermsSQL' :: ByteString -> DList StringSQL

ShowConstantTermsSQL Text
Constant SQL terms of Text.

Methods

showConstantTermsSQL' :: Text -> DList StringSQL

ShowConstantTermsSQL Text
Constant SQL terms of Text.

Methods

showConstantTermsSQL' :: Text -> DList StringSQL

ShowConstantTermsSQL LocalTime
Constant SQL terms of LocalTime.

Methods

showConstantTermsSQL' :: LocalTime -> DList StringSQL

ShowConstantTermsSQL ZonedTime
Constant SQL terms of ZonedTime. This generates ***NOT STANDARD*** SQL of TIMESTAMPTZ literal.

Methods

showConstantTermsSQL' :: ZonedTime -> DList StringSQL

ShowConstantTermsSQL TimeOfDay
Constant SQL terms of TimeOfDay.

Methods

showConstantTermsSQL' :: TimeOfDay -> DList StringSQL

ShowConstantTermsSQL UTCTime
Constant SQL terms of UTCTime. This generates ***NOT STANDARD*** SQL of TIMESTAMPTZ literal with UTC timezone.

Methods

showConstantTermsSQL' :: UTCTime -> DList StringSQL

ShowConstantTermsSQL Day
Constant SQL terms of Day.

Methods

showConstantTermsSQL' :: Day -> DList StringSQL

(PersistableWidth a, ShowConstantTermsSQL a) => ShowConstantTermsSQL (Maybe a)
Constant SQL terms of Maybe type. Width inference is required.

Methods

showConstantTermsSQL' :: Maybe a -> DList StringSQL