Database.Relational.Query.Relation

Description

This module defines re-usable Relation type to compose complex query.

Relation type

table :: Table r -> Relation () r

Simple Relation from Table.

derivedRelation :: TableDerivable r => Relation () r

Inferred Relation.

tableOf :: TableDerivable r => Relation () r -> Table r

Interface to derive Table type object.

relation :: QuerySimple (Projection Flat r) -> Relation () r

Finalize QuerySimple monad and generate Relation.

relation' :: SimpleQuery p r -> Relation p r

Finalize QuerySimple monad and generate Relation with place-holder parameter p.

aggregateRelation :: QueryAggregate (Projection Aggregated r) -> Relation () r

Finalize QueryAggregate monad and geneate Relation.

aggregateRelation' :: AggregatedQuery p r -> Relation p r

Finalize QueryAggregate monad and geneate Relation with place-holder parameter p.

data UniqueRelation p c r

Unique relation type to compose scalar queries.

unsafeUnique :: Relation p r -> UniqueRelation p c r

Unsafely specify unique relation.

unUnique :: UniqueRelation p c r -> Relation p r

Discard unique attribute.

uniqueRelation' :: QueryUnique (PlaceHolders p, Projection c r) -> UniqueRelation p c r

Finalize QueryUnique monad and generate UniqueRelation.

aggregatedUnique :: Relation ph r -> Pi r a -> (Projection Flat a -> Projection Aggregated b) -> UniqueRelation ph Flat b

Aggregated UniqueRelation.

Query using relation

query :: (MonadQualify ConfigureQuery m, MonadQuery m) => Relation () r -> m (Projection Flat r)

Join sub-query. Query result is not Maybe.

queryMaybe :: (MonadQualify ConfigureQuery m, MonadQuery m) => Relation () r -> m (Projection Flat (Maybe r))

Join sub-query. Query result is Maybe. The combinations of query and queryMaybe express inner joins, left outer joins, right outer joins, and full outer joins. Here is an example of a right outer join:

  outerJoin = relation $ do
    e <- queryMaybe employee
    d <- query department
    on $ e ?! E.deptId' .=. just (d ! D.deptId')
    return $ (,) |$| e |*| d

queryList :: MonadQualify ConfigureQuery m => Relation () r -> m (ListProjection (Projection c) r)

List sub-query, for IN and EXIST.

queryList' :: MonadQualify ConfigureQuery m => Relation p r -> m (PlaceHolders p, ListProjection (Projection c) r)

List sub-query, for IN and EXIST with place-holder parameter p.

queryScalar :: (MonadQualify ConfigureQuery m, ScalarDegree r) => UniqueRelation () c r -> m (Projection c (Maybe r))

Scalar sub-query.

queryScalar' :: (MonadQualify ConfigureQuery m, ScalarDegree r) => UniqueRelation p c r -> m (PlaceHolders p, Projection c (Maybe r))

Scalar sub-query with place-holder parameter p.

uniqueQuery' :: UniqueRelation p c r -> QueryUnique (PlaceHolders p, Projection c r)

Join unique sub-query with place-holder parameter p.

uniqueQueryMaybe' :: UniqueRelation p c r -> QueryUnique (PlaceHolders p, Projection c (Maybe r))

Join unique sub-query with place-holder parameter p. Query result is Maybe.

Direct style join

type JoinRestriction a b = Projection Flat a -> Projection Flat b -> Projection Flat (Maybe Bool)

Restriction predicate function type for direct style join operator, used on predicates of direct join style as follows.

  do xy <- query $
           relX inner relY on' [ x y -> ... ] -- this lambda form has JoinRestriction type
     ...

inner' infixl 8

Arguments

:: Relation pa a
Left query to join

-> Relation pb b
Right query to join

-> [JoinRestriction a b]
Join restrictions

-> Relation (pa, pb) (a, b)
Result joined relation

Direct inner join with place-holder parameters.

left' infixl 8

Arguments

:: Relation pa a
Left query to join

-> Relation pb b
Right query to join

-> [JoinRestriction a (Maybe b)]
Join restrictions

-> Relation (pa, pb) (a, Maybe b)
Result joined relation

Direct left outer join with place-holder parameters.

right' infixl 8

Arguments

:: Relation pa a
Left query to join

-> Relation pb b
Right query to join

-> [JoinRestriction (Maybe a) b]
Join restrictions

-> Relation (pa, pb) (Maybe a, b)
Result joined relation

Direct right outer join with place-holder parameters.

full' infixl 8

Arguments

:: Relation pa a
Left query to join

-> Relation pb b
Right query to join

-> [JoinRestriction (Maybe a) (Maybe b)]
Join restrictions

-> Relation (pa, pb) (Maybe a, Maybe b)
Result joined relation

Direct full outer join with place-holder parameters.

inner infixl 8

Arguments

:: Relation () a
Left query to join

-> Relation () b
Right query to join

-> [JoinRestriction a b]
Join restrictions

-> Relation () (a, b)
Result joined relation

Direct inner join.

left infixl 8

Arguments

:: Relation () a
Left query to join

-> Relation () b
Right query to join

-> [JoinRestriction a (Maybe b)]
Join restrictions

-> Relation () (a, Maybe b)
Result joined relation

Direct left outer join.

right infixl 8

Arguments

:: Relation () a
Left query to join

-> Relation () b
Right query to join

-> [JoinRestriction (Maybe a) b]
Join restrictions

-> Relation () (Maybe a, b)
Result joined relation

Direct right outer join.

full infixl 8

Arguments

:: Relation () a
Left query to join

-> Relation () b
Right query to join

-> [JoinRestriction (Maybe a) (Maybe b)]
Join restrictions

-> Relation () (Maybe a, Maybe b)
Result joined relation

Direct full outer join.

on' :: ([JoinRestriction a b] -> Relation pc (a, b)) -> [JoinRestriction a b] -> Relation pc (a, b) infixl 8

Apply restriction for direct join style.

Relation append

union :: Relation () a -> Relation () a -> Relation () a infixl 7

Union of two relations.

except :: Relation () a -> Relation () a -> Relation () a infixl 7

Subtraction of two relations.

intersect :: Relation () a -> Relation () a -> Relation () a infixl 8

Intersection of two relations.

unionAll :: Relation () a -> Relation () a -> Relation () a infixl 7

Union of two relations. Not distinct.

exceptAll :: Relation () a -> Relation () a -> Relation () a infixl 7

Subtraction of two relations. Not distinct.

intersectAll :: Relation () a -> Relation () a -> Relation () a infixl 8

Intersection of two relations. Not distinct.

union' :: Relation p a -> Relation q a -> Relation (p, q) a infixl 7

Union of two relations with place-holder parameters.

except' :: Relation p a -> Relation q a -> Relation (p, q) a infixl 7

Subtraction of two relations with place-holder parameters.

intersect' :: Relation p a -> Relation q a -> Relation (p, q) a infixl 8

Intersection of two relations with place-holder parameters.

unionAll' :: Relation p a -> Relation q a -> Relation (p, q) a infixl 7

Union of two relations with place-holder parameters. Not distinct.

exceptAll' :: Relation p a -> Relation q a -> Relation (p, q) a infixl 7

Subtraction of two relations with place-holder parameters. Not distinct.

intersectAll' :: Relation p a -> Relation q a -> Relation (p, q) a infixl 8

Intersection of two relations with place-holder parameters. Not distinct.