Database.Relational.Query.Type

Description

This module defines typed SQL.

Typed query statement

newtype Query p a

Query type with place-holder parameter p and query result type a.

Constructors

Query

Fields

untypeQuery :: String

unsafeTypedQuery

Arguments

:: String
Query SQL to type

-> Query p a
Typed result

Unsafely make typed Query from SQL string.

relationalQuery' :: Relation p r -> QuerySuffix -> Query p r

From Relation into typed Query with suffix SQL words.

relationalQuery :: Relation p r -> Query p r

From Relation into typed Query.

relationalQuerySQL :: Config -> Relation p r -> QuerySuffix -> String

From Relation into untyped SQL query string.

Typed update statement

data KeyUpdate p a

Update type with key type p and update record type a. Columns to update are record columns other than key columns, So place-holder parameter type is the same as record type a.

Constructors

KeyUpdate

Fields

updateKey :: Pi a p

untypeKeyUpdate :: String

unsafeTypedKeyUpdate :: Pi a p -> String -> KeyUpdate p a

Unsafely make typed KeyUpdate from SQL string.

typedKeyUpdate :: Table a -> Pi a p -> KeyUpdate p a

Make typed KeyUpdate from Table and key columns selector Pi.

typedKeyUpdateTable :: TableDerivable r => Relation () r -> Pi r p -> KeyUpdate p r

Make typed KeyUpdate object using derived info specified by Relation type.

derivedKeyUpdate :: TableDerivable r => Pi r p -> KeyUpdate p r

Make typed KeyUpdate from derived table and key columns selector Pi.

newtype Update p

Update type with place-holder parameter p.

Constructors

Update

Fields

untypeUpdate :: String

unsafeTypedUpdate :: String -> Update p

Unsafely make typed Update from SQL string.

typedUpdate' :: Config -> Table r -> UpdateTarget p r -> Update p

Make typed Update from Config, Table and UpdateTarget.

typedUpdate :: Table r -> UpdateTarget p r -> Update p

Make typed Update using defaultConfig, Table and UpdateTarget.

derivedUpdate' :: TableDerivable r => Config -> AssignStatement r (PlaceHolders p) -> Update p

Make typed Update from Config, derived table and AssignStatement

derivedUpdate :: TableDerivable r => AssignStatement r (PlaceHolders p) -> Update p

Make typed Update from defaultConfig, derived table and AssignStatement

typedUpdateAllColumn :: PersistableWidth r => Table r -> Restriction p r -> Update (r, p)

Make typed Update from Table and Restriction. Update target is all column.

derivedUpdateAllColumn' :: (PersistableWidth r, TableDerivable r) => Config -> RestrictedStatement r (PlaceHolders p) -> Update (r, p)

Make typed Update from Config, derived table and AssignStatement. Update target is all column.

derivedUpdateAllColumn :: (PersistableWidth r, TableDerivable r) => RestrictedStatement r (PlaceHolders p) -> Update (r, p)

Make typed Update from defaultConfig, derived table and AssignStatement. Update target is all column.

restrictedUpdateAllColumn

Arguments

:: PersistableWidth r

=> Table r

-> RestrictedStatement r (PlaceHolders p)
Restrict monad context

-> Update (r, p)

Deprecated: Use derivedUpdateAllColumn or typedUpdateAllColumn instead of this.

Directly make typed Update from Table and Restrict monad context. Update target is all column.

restrictedUpdateTableAllColumn :: (PersistableWidth r, TableDerivable r) => Relation () r -> RestrictedStatement r (PlaceHolders p) -> Update (r, p)

Deprecated: Use derivedUpdateAllColumn or typedUpdateAllColumn instead of this.

Directly make typed Update from Table and Restrict monad context. Update target is all column.

updateSQL :: Config -> Table r -> UpdateTarget p r -> String

Make untyped update SQL string from Table and UpdateTarget.

Typed insert statement

data Insert a

Insert type to insert record type a.

Constructors

Insert

Fields

untypeInsert :: String

chunkedInsert :: Maybe (String, Int)

untypeChunkInsert :: Insert a -> String

Statement to use chunked insert

chunkSizeOfInsert :: Insert a -> Int

Size to use chunked insert

unsafeTypedInsert' :: String -> String -> Int -> Insert a

Unsafely make typed Insert from single insert and chunked insert SQL.

unsafeTypedInsert :: String -> Insert a

Unsafely make typed Insert from single insert SQL.

typedInsert' :: PersistableWidth r => Config -> Table r -> Pi r r' -> Insert r'

Make typed Insert from Table and columns selector Pi with configuration parameter.

typedInsert :: PersistableWidth r => Table r -> Pi r r' -> Insert r'

Make typed Insert from Table and columns selector Pi.

derivedInsert :: (PersistableWidth r, TableDerivable r) => Pi r r' -> Insert r'

Table type inferred Insert.

typedInsertValue' :: Config -> Table r -> InsertTarget p r -> Insert p

Make typed Insert from Config, Table and monadic builded InsertTarget object.

typedInsertValue :: Table r -> InsertTarget p r -> Insert p

Make typed Insert from Table and monadic builded InsertTarget object.

derivedInsertValue' :: TableDerivable r => Config -> Register r (PlaceHolders p) -> Insert p

Make typed Insert from Config, derived table and monadic builded Register object.

derivedInsertValue :: TableDerivable r => Register r (PlaceHolders p) -> Insert p

Make typed Insert from defaultConfig, derived table and monadic builded Register object.

newtype InsertQuery p

InsertQuery type.

Constructors

InsertQuery

Fields

untypeInsertQuery :: String

unsafeTypedInsertQuery :: String -> InsertQuery p

Unsafely make typed InsertQuery from SQL string.

typedInsertQuery' :: Config -> Table r -> Pi r r' -> Relation p r' -> InsertQuery p

Make typed InsertQuery from columns selector Table, Pi and Relation with configuration parameter.

typedInsertQuery :: Table r -> Pi r r' -> Relation p r' -> InsertQuery p

Make typed InsertQuery from columns selector Table, Pi and Relation.

derivedInsertQuery :: TableDerivable r => Pi r r' -> Relation p r' -> InsertQuery p

Table type inferred InsertQuery.

insertQuerySQL :: Config -> Table r -> Pi r r' -> Relation p r' -> String

Make untyped insert select SQL string from Table, Pi and Relation.

Typed delete statement

newtype Delete p

Delete type with place-holder parameter p.

Constructors

Delete

Fields

untypeDelete :: String

unsafeTypedDelete :: String -> Delete p

Unsafely make typed Delete from SQL string.

typedDelete' :: Config -> Table r -> Restriction p r -> Delete p

Make typed Delete from Config, Table and Restriction.

typedDelete :: Table r -> Restriction p r -> Delete p

Make typed Delete from Table and Restriction.

derivedDelete' :: TableDerivable r => Config -> RestrictedStatement r (PlaceHolders p) -> Delete p

Make typed Delete from Config, derived table and RestrictContext

derivedDelete :: TableDerivable r => RestrictedStatement r (PlaceHolders p) -> Delete p

Make typed Delete from defaultConfig, derived table and RestrictContext

deleteSQL :: Config -> Table r -> Restriction p r -> String

Make untyped delete SQL string from Table and Restriction.

Generalized interfaces

class UntypeableNoFetch s where

Untype interface for typed no-result type statments with single type parameter which represents place-holder parameter p.

Minimal complete definition

untypeNoFetch

Methods

untypeNoFetch :: s p -> String